+++ package.json
@@ -20,6 +20,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-is": "^18.3.1",
+    "sass": "^1.77.8",
     "typescript-eslint": "^8.1.0",
     "vite-plugin-style-import": "^2.0.0",
     "zustand": "^4.5.5"
@@ -42,8 +43,6 @@
     "eslint-plugin-react-refresh": "^0.4.7",
     "globals": "^15.9.0",
     "husky": "^9.1.4",
-    "less": "^4.2.0",
-    "less-loader": "^12.2.0",
     "lint-staged": "^15.2.9",
     "prettier": "^3.3.3",
     "typescript": "^5.2.2",
