--- shared/config/colors.ts
@@ -1,16 +1,0 @@
-export const colors = {
-  token: {
-    primaryBlue: '#514EF3',
-    primaryNavy: '#092C4C',
-    black: '#000000',
-    lightGrey10: '#F6FAFD', //светло-серый10
-    grayLightBg20: '#EEF6FB', //серый20
-    grayDark: '#7E92A2', // темно-серый70
-    grayMedium30: '#EAEEF4', // средне-серый30
-    lightBlue: '#ECECFE',
-    white: '#FFFFFF',
-    green: '#2DC8A8',
-    pink: '#FE8084',
-    yellow: '#FFC357',
-  },
-}
