+++ src/App.tsx
@@ -4,6 +4,7 @@ function App() {
   return (
     <div className="App">
       <Button type="primary">Button</Button>
+      <Button type="default">Button</Button>
     </div>
   )
 }
