+++ src/shared/config/theme.ts
@@ -1,0 +1,21 @@
+import { ThemeConfig } from 'antd'
+import { Colors } from './colors.ts'
+
+export const theme: ThemeConfig = {
+  token: {
+    colorPrimary: Colors.primaryBlue,
+    colorBgBase: Colors.lightGrey10,
+    colorPrimaryText: Colors.primaryNavy,
+    colorInfoBg: Colors.grayLightBg20,
+    colorIcon: Colors.grayDark,
+    colorBorder: Colors.grayMedium30,
+    colorBgContainerDisabled: Colors.lightBlue,
+    colorBgContainer: Colors.white,
+    colorFill: Colors.green,
+    colorError: Colors.pink,
+    colorWarning: Colors.yellow,
+  },
+  components: {
+    Button: {},
+  },
+}
